requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# Redis
redis>=5.0.1

# Auth
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0

# Settings & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utils
python-multipart>=0.0.6
